Provided evc_parse.{h,c} to enable EVC NALU parsing code sharing between EVC parser and frame_merge bsf
